+++ packages/create-content-sdk-app/src/templates/nextjs-app-router (beta)/next.config.ts
@@ -19,6 +19,21 @@ const nextConfig: NextConfig = {
       },
     ],
   },
+  // use this configuration to serve the sitemap.xml and robots.txt files from the API route handlers
+  rewrites: async () => {
+    return [
+      {
+        source: '/sitemap:id([\\w-]{0,}).xml',
+        destination: '/api/sitemap',
+        locale: false,
+      },
+      {
+        source: '/robots.txt',
+        destination: '/api/robots',
+        locale: false,
+      },
+    ];
+  },
 };
 
 const withNextIntl = createNextIntlPlugin();